/utils: renamed files to use names valid on Windows
Closes #502
