+++ packages/tools/utils/index.js
@@ -3,13 +3,11 @@ export * from './src/modules/[any-any]-[iterable-object]'
 export * from './src/modules/[any-any]-[object-array]'
 export * from './src/modules/[any-any]-[object-number]'
 export * from './src/modules/[any-any]-[object-object]'
-export * from './src/modules/[any-any]:accumcb-[array-any]'
 export * from './src/modules/[any-array]-[any-array]'
 export * from './src/modules/[any-array]-[array-object]'
 export * from './src/modules/[any-boolean]-[array-array]'
 export * from './src/modules/[any-boolean]-[object-any]'
 export * from './src/modules/[any-number]-[array-number]'
-export * from './src/modules/[any-number]:accumcb-[array-number]'
 export * from './src/modules/[any-object]-[array-object]'
 export * from './src/modules/[string-string]-[object-object]'
 export * from './src/modules/any-[any-boolean]'
@@ -41,7 +39,7 @@ export * from './src/modules/array-number'
 export * from './src/modules/array-object'
 export * from './src/modules/array-string'
 export * from './src/modules/buffer-any'
-export * from './src/modules/constructor-[[any-any]:accumcb-[array-any]]'
+export * from './src/modules/constructor-[reduceCb[any-any]-[array-any]]'
 export * from './src/modules/iterable-boolean'
 export * from './src/modules/iterable-number'
 export * from './src/modules/iterable-object'
@@ -55,6 +53,7 @@ export * from './src/modules/object-array'
 export * from './src/modules/object-boolean'
 export * from './src/modules/object-number'
 export * from './src/modules/object-object'
+export * from './src/modules/reduceCb[any-any]-[array-any]'
 export * from './src/modules/string_proto-array'
 export * from './src/modules/string_proto-boolean'
 export * from './src/modules/string_proto-string'
