--- packages/tools/utils/src/modules/[any-number]:accumcb-[array-number].js
@@ -1,59 +1,0 @@
-/**
-* @module @svizzle/utils/[any-number]:accumcb-[array-number]
-*/
-
-import * as _ from 'lamb';
-
-/**
- * Return a function expecting an array of objects and returning the max of results
- * of applying the provided fuction on all of the array items
- * @see
- {@link module:@svizzle/utils/array-number.arrayMax|arrayMax},
- {@link module:@svizzle/utils/string-[array-number].arrayMaxBy|arrayMaxBy}
- *
- * @function
- * @arg {function} fn
- * @return {function} - Array -> Number
- *
- * @example
-
-> maxWithAbsSin = arrayMaxWith(_.pipe([Math.sin, Math.abs]))
-> maxWithAbsSin([-Math.PI/2, -Math.PI/4])
-1
-> maxWithAbsSin([Math.PI/4, Math.PI/6])
-0.7071067811865475
-
- *
- * @since 0.1.0
- */
-export const arrayMaxWith = fn => _.reduceWith((max, item) => {
-	const value = fn(item);
-
-	return value > max ? value : max;
-}, -Infinity);
-
-/**
- * Return a function expecting an array of objects and returning the min of results
- * of applying the provided fuction on all of the array items
- * @see
- {@link module:@svizzle/utils/array-number.arrayMin|arrayMin},
- {@link module:@svizzle/utils/string-[array-number].arrayMinBy|arrayMinBy}
- *
- * @function
- * @arg {function} fn
- * @return {function} - Array -> Number
- *
- * @example
-> minWithAbsSin = arrayMinWith(_.pipe([Math.sin, Math.abs]))
-> minWithAbsSin([-Math.PI/2, -Math.PI/4])
-0.7071067811865475
-> minWithAbsSin([Math.PI/4, Math.PI/6])
-0.49999999999999994
- *
- * @since 0.1.0
- */
-export const arrayMinWith = fn => _.reduceWith((min, item) => {
-	const value = fn(item);
-
-	return value < min ? value : min;
-}, Infinity);
