+++ packages/tools/utils/src/modules/array-number.js
@@ -28,7 +28,7 @@ export const makeRandomNumInRange = ([min, max]) =>
  * Return the max of the numbers in the provided array
  * @see
  {@link module:@svizzle/utils/string-[array-number].arrayMaxBy|arrayMaxBy},
- {@link module:@svizzle/utils/[any-number]:accumcb-[array-number].arrayMaxWith|arrayMaxWith}
+ {@link module:@svizzle/utils/[any-number]-[array-number].arrayMaxWith|arrayMaxWith}
  *
  * @function
  * @arg {array} array
@@ -46,7 +46,7 @@ export const arrayMax = _.apply(Math.max);
  * Return the min of the numbers in the provided array
  * @see
  {@link module:@svizzle/utils/string-[array-number].arrayMinBy|arrayMinBy},
- {@link module:@svizzle/utils/[any-number]:accumcb-[array-number].arrayMinWith|arrayMinWith}
+ {@link module:@svizzle/utils/[any-number]-[array-number].arrayMinWith|arrayMinWith}
  *
  * @function
  * @arg {array} array
