+++ packages/tools/utils/src/modules/itercb[any-any]-[iterable-array].md
