+++ packages/tools/utils/src/modules/itercb[any-any]-[iterable-object].md
