+++ packages/tools/utils/src/modules/itercb[any-boolean]-[iterable-number].md
